feat 分账接收方功能

+++ daxpay-single/daxpay-single-service/src/main/java/org/dromara/daxpay/service/entity/order/allocation/AllocOrderDetail.java
@@ -31,8 +31,8 @@ public class AllocOrderDetail extends MchAppBaseEntity {
     /** 分账接收方编号 */
     private String receiverNo;
 
-    /** 分账比例 */
-    private Integer rate;
+    /** 分账比例(百分之多少) */
+    private BigDecimal rate;
 
     /** 分账金额 */
     private BigDecimal amount;
